Added decrement inrement and deletion.

+++ pkg/models/produckt.go
@@ -3,5 +3,5 @@ package models
 type Product struct {
   UserID string
   ProductId string
-  Amount int
+  Amount uint8
 }
